scripts: Add script to generate strace entries
This patch introduces a python script that is meant to
update the file 'strace.list' with missing strace entries.
The file is update using a 'syscall_nr.h' from a linux-user
target build folder. When running the script, a target arch
can be specified from the command line. If no target is specified,
the 'strace.list' is update using the first available 'syscall_nr.h'
file from the first available target build folder.

Signed-off-by: Filip Bozuta <Filip.Bozuta@syrmia.com>